+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
  - Re-ordered Devices/Ports/Panels in Room configuration for consistency
  - Better combine state tracking of rooms
+ 
+### Removed
+ - Removed AudioDestinations collection on the Room
 
 ## [9.0.0] - 2018-11-08
 ### Changed